+++ Atlas/Net/AtlasXMLEncoder.cpp
@@ -116,7 +116,7 @@ void AXMLEncoder::walkTree(int nest, string name, const AObject& list)
 	}
 	if (list.isInt()) {
 		printf("%s<int%s>%li</int>\n",
-			pre.c_str(), nam.c_str(),list.asLong()
+			pre.c_str(), nam.c_str(),list.asInt()
 		);
 	}
 	if (list.isLong()) {
